+++ R/get_object.R
@@ -4,12 +4,21 @@
 #' @template object
 #' @template bucket
 #' @param file An R connection, or file name specifying the local file to save the object into.
+#' @param request_body For \code{select_object}, an XML request body as described in the \href{https://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectSELECTContent.html}{SELECT API documentation}.
 #' @param headers List of request headers for the REST call.
 #' @param parse_response Passed through to \code{\link{s3HTTP}}, as this function requires a non-default setting. There is probably no reason to ever change this.
 #' @template dots
-#' @details \code{get_object} retrieves an object into memory as a raw vector. \code{save_object} saves an object to a local file. By contrast, \code{\link{head_object}} checks whether an object exists by executing an HTTP HEAD request; this can be useful for checking object headers such as \dQuote{content-length} or \dQuote{content-type}.
+#' @details \code{get_object} retrieves an object into memory as a raw vector. This page describes \code{get_object} and several wrappers that provide additional useful functionality.
+#' 
+#' \code{save_object} saves an object to a local file without bringing it into memory.
+#' 
+#' \code{s3connection} provides a \code{\link[base]{connection}} interface to an S3 object.
+#' 
+#' \code{select_object} uses the \href{https://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectSELECTContent.html}{SELECT API} to select part of a CSV or JSON object. This requires constructing and passing a fairly tedious request body, which users will have to construct themselves according to the documentation.
 #'
 #' Some users may find the raw vector response format of \code{get_object} unfamiliar. The object will also carry attributes, including \dQuote{content-type}, which may be useful for deciding how to subsequently process the vector. Two common strategies are as follows. For text content types, running \code{\link[base]{charToRaw}} may be the most useful first step to make the response human-readable. Alternatively, converting the raw vector into a connection using \code{\link[base]{rawConnection}} may also be useful, as that can often then be passed to parsing functions just like a file connection would be.
+#'
+#' Higher-level functions
 #' 
 #' @examples
 #' \dontrun{
@@ -35,11 +44,29 @@
 #'   # return parts of an object
 #'   ## use 'Range' header to specify bytes
 #'   get_object(object = obj[[1]], headers = list('Range' = 'bytes=1-120'))
+#'  
+#'   # example of streaming connection
+#'   ## setup a bucket and object
+#'   b <- put_bucket("myexamplebucket")
+#'   s3write_using(mtcars, bucket = b, object = "mtcars.csv", FUN = utils::write.csv)
+#'   
+#'   ## setup the connection
+#'   con <- s3connection("mtcars.csv", bucket = b)
+#'   
+#'   ## line-by-line read
+#'   while(length(x <- readLines(con, n = 1L))) {
+#'     print(x)
+#'   }
+#'   ## cleanup
+#'   close(con)
+#'   delete_bucket("myexamplebucket")
 #' }
 #' @return If \code{file = NULL}, a raw object. Otherwise, a character string containing the file name that the object is saved to.
-#' @references \href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html}{API Documentation: GET Object}
-#' @references \href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html}{API Documentation: GET Object torrent}
-#' @seealso \code{\link{get_bucket}}, \code{\link{head_object}}, \code{\link{put_object}}, \code{\link{delete_object}}
+#' @references
+#'  \href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html}{API Documentation: GET Object}
+#'  \href{http://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectGET.html}{API Documentation: GET Object torrent}
+#'  \href{https://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectSELECTContent.html}{API Documentation: SELECT Object}
+#' @seealso \code{\link{get_bucket}}, \code{\link{object_exists}}, \code{\link{head_object}}, \code{\link{put_object}}, \code{\link{delete_object}}
 #' @export
 get_object <- 
 function(object, 
@@ -92,6 +119,34 @@ function(object,
     return(file)
 }
 
+#' @rdname get_object
+#' @export
+select_object <- 
+function(
+  object,
+  bucket,
+  request_body,
+  headers = list(),
+  parse_response = FALSE,
+  ...
+) {
+    if (missing(bucket)) {
+        bucket <- get_bucketname(object)
+    } 
+    object <- get_objectkey(object)
+    
+    r <- s3HTTP(verb = "POST", 
+                bucket = bucket,
+                path = paste0("/", object),
+                headers = headers,
+                query = list(select = "", "select-type" = "2"),
+                request_body = request_body,
+                parse_response = parse_response,
+                ...)
+    cont <- httr::content(r, as = "raw")
+    return(cont)
+}
+
 #' @title Get object torrent
 #' @description Retrieves a Bencoded dictionary (BitTorrent) for an object from an S3 bucket.
 #' 
@@ -113,4 +168,23 @@ get_torrent <- function(object, bucket, ...) {
                 query = list(torrent =""),
                 ...)
     return(content(r, "raw"))
+}
+
+#' @rdname get_object
+#' @export
+s3connection <-
+function(object,
+         bucket,
+         headers = list(),
+         ...) {
+    if (missing(bucket)) {
+        bucket <- get_bucketname(object)
+    } 
+    object <- get_objectkey(object)
+    r <- s3HTTP(verb = "connection",
+                bucket = bucket,
+                path = paste0("/", object),
+                headers = headers,
+                ...)
+    return(r)
 }